+++ src/app/feature/home/components/card/card.component.ts
@@ -2,8 +2,11 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { Character } from '@home/shared/model/character.model';
 import { CharacterService } from '@home/shared/services/character/character.service';
 import { ComicService } from '@home/shared/services/comic/comic.service';
+import { FavoriteService } from '@home/shared/services/favorite/favorite.service';
 import { ModalComponent } from '@shared/components/modal/modal.component';
+import { AlertService } from '@shared/services/alert.service';
 import { fromEvent } from 'rxjs';
+import { FavoriteComponent } from '../favorite/favorite.component';
 
 @Component({
   selector: 'app-card',
@@ -19,7 +22,9 @@ export class CardComponent implements OnInit {
 
   constructor(
     private apiCharacter: CharacterService,
-    private apiComic: ComicService
+    private apiComic: ComicService,
+    private apiFavorite: FavoriteService,
+    private alertService: AlertService,
   ) { }
 
   ngOnInit(): void {
@@ -49,6 +54,17 @@ export class CardComponent implements OnInit {
       this.record = comic;
       this.modal.open(this.record, true, true);
     });
+  }
+
+  add(data: any) {
+    console.log(data.id);
+    if(this.apiFavorite.add(data)) {
+      this.alertService.successAlert('The comic was added successfuly');
+      this.modal.close();
+    } else {
+      this.alertService.infoAlert('The comic can\'t added beacause it\'s already in the list')
+    }
+
   }
 
   doSearch() {