+++ src/app/feature/home/components/favorite/favorite.component.ts
@@ -18,7 +18,6 @@ export class FavoriteComponent implements OnInit {
   constructor(
     private apiComics: ComicService,
     private apiFavorite: FavoriteService,
-    private alertService: AlertService,
   ) { }
 
   ngOnInit(): void {
@@ -40,17 +39,6 @@ export class FavoriteComponent implements OnInit {
       const [record] = this.favorites.filter((e: {id: number}) => e.id == id);
       console.log(record);
       this.modal.open(record, true, false);
-  }
-
-  add(data: any) {
-    if(this.apiFavorite.add(data)) {
-      this.alertService.successAlert('The comic was added successfuly');
-      this.get();
-      this.modal.close();
-    } else {
-      this.alertService.infoAlert('The comic can\'t added beacause it\'s already in the list')
-    }
-
   }
 
   delete(id:number) {